# server/api.yaml
openapi: 3.0.0
info:
  title: JobSeeker API
  description: Dat Ngoc Ngan Ba Nghia.
  version: 0.1.9
servers:
  - url: http://localhost:8000
    description: My local server.
  - url: https://job-seeker-smy5.onrender.com
    description: Server is deloyed.
...
paths:
...
  /company/list:
    get:
      tags:
        - Company
      summary: Get Company
      responses:
        '200':
          description: 'Successfully.'
        '500':
          description: 'Failed'
...
components:
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header 